Delivery version commit
Rework cutting mechanic later. This is a good demo for the game in vr, the only problem is that the cutting is not stable enough, and fix the grabbing objets. Maybe create own system in VR for this. All frameworks used have a lot of issues regarding this. Creating own using as base SteamVR, maybe its good idea